# src/swagger/documentation.yaml
openapi: 3.0.0
info:
  title: My API
  description: API documentation for my application
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Section:
      type: object
      required:
         - name
      properties:
        id:
          type: integer
          description: Section ID
        name:
          type: string
          description: Section Name
      example:
          id: 1
          name: Popular Last Week
    Category:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          description: Category ID
        name:
          type: string
          description: Category Name
        description:
          type: string
          description: Category's Description
        parent:
          type: integer
          description: Category's Parent Category (Can be Null)
          nullable: true
        tier:
          type: integer
          description: Category's Tier (Can be Null)
          nullable: true
      example:
        id: 6
        name: Headphones
        description: Audio output device, can be connected to PC, Laptop, Phone, Tablet.
        parent: 2
        tier: 2
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
      example:
        message: Resource not found
  responses:
      NotFoundError:
        description: Resource not found
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrorResponse'
...
servers:
  - url: http://localhost:3000
    description: Local server
...
paths:
...
  /api/sections:
    get:
      summary: Retrieve all sections
      description: Returns a list of sections.
      tags: [Sections]
      responses:
        200:
          description: A list of sections.
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Section'
        404:
           $ref: '#/components/responses/NotFoundError'
    post:
      summary: Create a new section
      description: Creates a new section.
      tags: [Sections]
      parameters:
        - in: body
          name: name
          required: true
          schema:
            type: istring
      security:
        - bearerAuth: []
      responses:
        201:
          description: Section created successfully.
  /api/sections/{id}:
    get:
      summary: Retrieve a section by ID
      description: Returns a section object.
      tags: [Sections]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: A section object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Section'
        404:
           $ref: '#/components/responses/NotFoundError'
  /api/sections/{id}/categories:
    get:
      summary: Retrieve categories for a section
      description: Returns a list of categories which are bound to the Section of ID.
      tags: [Sections, Categories]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: A list of categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        400:
          $ref: '#/components/responses/NotFoundError'
  /api/sections/{id}/:
    patch:
      summary: Update a section
      tags: [Sections]
      description: Updates a section.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: body
          name: name
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Section updated successfully.
    delete:
      summary: Delete a section
      tags: [Sections]
      description: Deletes a section.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        204:
          description: Section deleted successfully.
  /api/sections/{id}/categories/:
    post:
      summary: Add a category to a section
      description: Adds a category to a section.
      tags: [Sections, Categories]
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: sectionId
          required: true
          schema:
            type: integer
        - in: body
          name: categoryId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Category added successfully.
  /api/sections/{id}/categories/{categoryId}/:
    delete:
      summary: Remove a category from a section
      description: Removes a category from a section.
      tags: [Sections, Categories]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Category removed successfully.

#------------------------------ PRODUCTS      
...